backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
google-generativeai
Pillow
dj-database-url
gunicorn
scikit-learn
joblib
transformers
torch
sentencepiece
